+++ 02makeBootGcc.sh
@@ -9,7 +9,9 @@ fi
    --with-gnu-ld --disable-shared --enable-languages=c \
    --disable-werror )
 # keep build quiet so we can see any stderr reports.
-cat quiet bootgcc_build/Makefile > Makefile
-mv Makefile bootgcc_build/Makefile
+if test -f quiet; then
+    cat quiet bootgcc_build/Makefile > Makefile
+    mv Makefile bootgcc_build/Makefile
+fi
 # note, make.log contains the stderr output of the build.
 (cd bootgcc_build ; make all-gcc install-gcc 2>&1 | tee $BUILDSOURCES/make.log)
