+++ 03makeNewLib.sh
@@ -11,7 +11,9 @@ rm -rf $PREFIX/newlib_build/*
     --with-gnu-ld --with-gnu-as --disable-newlib-io-float \
    --disable-werror )
 # keep build quiet so we can see any stderr reports.
-cat quiet $PREFIX/newlib_build/Makefile > $BUILDSOURCES/Makefile
-mv $BUILDSOURCES/Makefile $PREFIX/newlib_build/Makefile
+if test -f quiet; then
+    cat quiet $PREFIX/newlib_build/Makefile > $BUILDSOURCES/Makefile
+    mv $BUILDSOURCES/Makefile $PREFIX/newlib_build/Makefile
+fi
 # note, make.log contains the stderr output of the build.
 (cd $PREFIX/newlib_build ;  make all install 2>&1 | tee $BUILDSOURCES/make.log)
